more unknown data chunks added
in larger files some data was being missed. it seems that there are even more undocumented chunks in the files. These have now been accounted for and it should work. If 'In Index Chunk' shows, then you have definitely got to the end of the file and read all the readable data.